Déplace le chiffrement Vault dans un adaptateur séparé
On a donc deux adaptateurs de chiffrement : un sachant faire du Vault, un ne faisant que retourner ce qu'il reçoit.
Avec la fabrique d'adaptateur, on peut ainsi déployer notre code sur un environnement où Vault n'est pas disponible.
