+++ src/adaptateurs/fabriqueAdaptateurChiffrement.js
@@ -1,8 +1,11 @@
 const { chiffrement } = require('./adaptateurEnvironnement');
 const adaptateurChiffrement = require('./adaptateurChiffrement');
+const adaptateurChiffrementVault = require('./adaptateurChiffrementVault');
 
 const fabriqueAdaptateurChiffrement = () => {
-  if (chiffrement().utiliseChiffrementVault()) return adaptateurChiffrement;
+  if (chiffrement().utiliseChiffrementVault())
+    return adaptateurChiffrementVault;
+
   return adaptateurChiffrement;
 };
